Finished the immutable refactor (with some corners cut that will need to be added back after the fact like lights.

+++ projects/marloth/shared/scenery/src/scenery/Depiction.kt
@@ -2,7 +2,7 @@ package scenery
 
 import mythic.breeze.Armature
 
-typealias Id = Int
+typealias Id = Long
 
 data class DepictionAnimation(
     val animationIndex: Int,
